File_Management_System_New/backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
bleach==6.0.0
certifi==2022.9.24
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-cors-headers==3.13.0
django-debug-toolbar==3.8.1
django-environ==0.9.0
django-filter==22.1
django-import-export==2.9.0
django-rest-auth==0.9.5
# django-rest-auth-forked==0.9.5
django-wkhtmltopdf==3.4.0
djangorestframework==3.14.0
drf-api-logger==1.1.12
drf-yasg==1.21.7
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
pandas==2.0.1
pdfkit==1.0.0
Pillow==9.3.0
psutil==5.9.5
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
redis==4.4.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tablib==3.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.13
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
mysqlclient==2.2.0
django-filter>=22.1,<24.0
django-pandas==0.6.6
opencv-python-headless==4.8.1.78
mtcnn==0.1.1
pip-chill==1.0.3
pysocks==1.7.1
tf-keras==2.17.0


